+++ packages/stream_chat/test/src/client/client_test.dart
@@ -1915,6 +1915,50 @@ void main() {
       verifyNoMoreInteractions(api.channel);
     });
 
+    test('`.addChannelMembers` with hideHistoryBefore', () async {
+      const channelType = 'test-channel-type';
+      const channelId = 'test-channel-id';
+      const channelCid = '$channelType:$channelId';
+
+      final members = List.generate(
+        3,
+        (index) => Member(userId: 'test-user-id-$index'),
+      );
+
+      final memberIds = members.map((e) => e.userId!).toList(growable: false);
+      final hideHistoryBefore = DateTime.parse('2024-01-01T00:00:00Z');
+
+      when(() => api.channel.addMembers(
+            channelId,
+            channelType,
+            memberIds,
+            hideHistoryBefore: hideHistoryBefore,
+          )).thenAnswer((_) async => AddMembersResponse()
+        ..channel = ChannelModel(cid: channelCid)
+        ..members = members);
+
+      final res = await client.addChannelMembers(
+        channelId,
+        channelType,
+        memberIds,
+        hideHistoryBefore: hideHistoryBefore,
+      );
+
+      expect(res, isNotNull);
+      expect(res.channel.cid, channelCid);
+      expect(res.members.length, memberIds.length);
+
+      verify(
+        () => api.channel.addMembers(
+          channelId,
+          channelType,
+          memberIds,
+          hideHistoryBefore: hideHistoryBefore,
+        ),
+      ).called(1);
+      verifyNoMoreInteractions(api.channel);
+    });
+
     test('`.removeChannelMembers`', () async {
       const channelType = 'test-channel-type';
       const channelId = 'test-channel-id';